Revert "Extend all to more resources"
